app: move body percussion under components
