+++ packages/app/src/components/body-percussion/common.ts
