+++ packages/app/src/components/body-percussion/intro.tsx
